requirements.txt
requests==2.32.*
flask==3.1.*
anthropic==0.49.*
pinecone==6.0.*
openai==1.66.*
retell-sdk==5.22.*
gunicorn==23.0.*
python-dotenv==1.1.*
